# cmd/agents-server/internal/docs/swagger.yaml
components:
  schemas:
...
    handler.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/protocol.APIError'
      type: object
...
    handler.projectDeleteResp:
      properties:
        deleted:
          type: boolean
        storage_error:
          type: string
      type: object
...
    handler.projectReq:
      properties:
        env:
          description: |-
            Env is the environment the project's container is created with;
            optional, and empty means none.
          items:
            $ref: '#/components/schemas/store.EnvVar'
          type: array
          uniqueItems: false
        name:
          type: string
        sandbox_id:
          description: SandboxID is what the project runs on — the machine and the
            image.
          type: string
      required:
      - name
      - sandbox_id
      type: object
    handler.projectUpdateReq:
      properties:
        env:
          items:
            $ref: '#/components/schemas/store.EnvVar'
          type: array
          uniqueItems: false
        name:
          type: string
        revision:
          type: integer
        sandbox_id:
          type: string
      required:
      - name
      - sandbox_id
      type: object
...
    handler.sandboxStateResp:
      properties:
        state:
          description: State is absent | stopped | running.
          type: string
      type: object
    handler.sandboxStopResp:
      properties:
        stopped:
          type: boolean
      type: object
...
    protocol.APIError:
      properties:
        code:
          type: string
        message:
          type: string
      type: object
...
    store.EnvVar:
      properties:
        key:
          type: string
        value:
          type: string
      type: object
...
    store.Project:
      properties:
        created_at:
          type: string
        id:
          type: string
        name:
          description: |-
            Name is display only — the storage is keyed by ID, so a rename moves
            nothing. Unique per (owner, sandbox) via idx_projects_owner_sandbox_name.
          type: string
        owner_id:
          type: string
        revision:
          description: |-
            Revision is the expected-revision CAS every update lands against.
            RuntimeGen is the workbench's ONE runtime axis: it moves when this
            project's own content changes AND when the sandbox it names changes
            underneath it, so the instance cache and the terminal registry need a
            single fence rather than one per entity. A rename moves neither
            container nor terminal.
          type: integer
        sandbox_id:
          description: |-
            SandboxID is what the project runs on. It may move only to a sandbox at
            the SAME type and destination (checkMove, else 409): the freeze is on
            which machine, not on the image, which edits freely on the sandbox row.
          type: string
        session_count:
          description: |-
            SessionCount is how many sessions bind this project — filled by List
            (scanonly), so a delete knows whether it will be refused.
          type: integer
        storage_hint:
          description: |-
            StorageHint names where the files live — the named volume on the
            sandbox's daemon. Derived per response by the handler for admins only,
            never stored: a delete DESTROYS that storage, so the UI can say what
            will be lost.
          type: string
        updated_at:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Value is "Bearer <token>" — the token printed at server startup
        (or set via --token).
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 'REST API of agents-server, the web server over the agents-go SDK:
    sessions, runs, agents, MCP servers, memories, settings, skills, providers, workflows,
    triggers, guardrails, sandboxes and projects. A run starts over REST (POST /sessions/{id}/runs)
    or the WebSocket protocol (docs/reference/protocol.md) and streams over either;
    this spec covers the HTTP surface.'
  title: agents-server API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /projects:
    get:
      description: Every row carries session_count. storage_hint (where the files
        live) is reported to admins only.
      parameters:
      - description: 'admin: every owner''s projects'
        in: query
        name: all
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.Project'
                type: array
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: all=true by a member
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.projectReq'
                description: Name and sandbox
                summary: project
        description: Name and sandbox
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.projectDetail'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: name already in use on this sandbox
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: 'Deletes the working tree too — the container and its volume are
        removed. The owner deletes their own; an admin deletes any. The row is gone
        whenever this answers 200: a storage_error means the STORAGE could not be
        reclaimed and is left for the operator, not that the project survived.'
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.projectDeleteResp'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: sessions still bound
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      description: 'The one endpoint that returns a project''s environment — names,
        with every value masked. Listings never carry it at all. Owner only: an environment
        is not part of an admin''s management reach.'
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.projectDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
    put:
      description: A value sent back as its mask keeps what is stored; any other value
        replaces it. An environment change replaces the project's container at its
        next run and severs its terminals; a rename does none of that.
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.projectUpdateReq'
                description: Name and environment
                summary: project
        description: Name and environment
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.projectDetail'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: name already in use, or the project changed concurrently —
            re-read and retry
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /projects/{id}/export:
    get:
      description: 'Streams /workspace as an uncompressed tar. Owner only, and audited:
        this takes the whole tree off the machine.'
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/x-tar:
              schema:
                type: string
          description: tar stream
        "404":
          content:
            application/x-tar:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "502":
          content:
            application/x-tar:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Export the project's working tree
      tags:
      - projects
  /projects/{id}/sandbox:
    get:
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.sandboxStateResp'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Project sandbox status
      tags:
      - projects
  /projects/{id}/sandbox/rebuild:
    post:
      description: 'Discards the container and creates a fresh one from the current
        sandbox and environment. Files under /workspace survive; anything installed
        into the container does not, and commands running in it fail. Synchronous.
        Owner or admin. Refused on a sandbox whose instance IS the storage (E2B-compatible):
        export first.'
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: rebuilt
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Rebuild the project's sandbox
      tags:
      - projects
  /projects/{id}/sandbox/start:
    post:
      description: Synchronous, and can take an image pull's worth of time. Owner
        or admin.
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: running
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Start the project's sandbox
      tags:
      - projects
  /projects/{id}/sandbox/stop:
    post:
      description: 'Keeps the working tree. `stopped: false` means a run or terminal
        is still using it and the stop happens when that ends. Owner or admin.'
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.sandboxStopResp'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Stop the project's sandbox
      tags:
      - projects
...
servers:
- url: /api/v1